Add the ability to load different IPTS firmware per SKU
All calls that load firmware inside the IPTS driver are wrapped with the new
ipts_request_firmware function. It requests the model name and SKU from DMI and
then tries to load the firmware from the following locations:

 - intel/ipts/[MODEL NAME]/[MODEL SKU]
 - intel/ipts/[MODEL NAME]
 - intel/ipts

So for my SB2 13" that would result in:

 - intel/ipts/Surface_Book_2/Surface_Book_1832
 - intel/ipts/Surface_Book_2
 - intel/ipts

Since `intel/ipts` was the directory that was checked before, merging this
would not influence existing installations.

However, for new installations, it would be possible to install the firmware
blobs for all models without a check at setup. This also makes it possible to
package the firmware files and install them through the package manager.

Signed-off-by: Dorian Stoll <dorian.stoll@tmsp.io>