pkg/ubuntu: Disable atomisp
While the atomisp driver might be used for Surface 3 and Surface Pro 3
cameras in the future, it currently resides in staging and doesn't seem
to work (at least according to its TODO file). It furthermore causes
a notice to be printed, warning about the use of trace_printk(). So
disable it for now to shut up that warning.