Remove setup script
This script is essentially useless and was only kept to help transition
users from the old JakeDay repository. All it does now is, if asked, set
the hardware clock to local time (which is also described in the wiki)
and print some (partially outdated) instructions.

Since the instructions are referencing outdated kernel versions, just
remove the script entirely instead of bothering to update them.