pkg: fedora: Add support for boot-on-btrfs
When the /boot directory is not its own partition but a subvolume or
a subdirectory on a BTRFS partition, then the path that GRUB uses to
boot the kernel is different.

For example instead of /boot/vmlinuz-6.4.12-1.surface.fc38.x86_64 it
will now look like /@boot/vmlinuz-6.4.12-1.surface.fc38.x86_64.

grubby expects the path that is noted inside of the BLS file, so we
need to find the corresponding file and extact the correct path.