pkg: fedora: Disable wrong NX compatibility flag
On Fedora the kernel needs to be loaded into RWX memory by GRUB.
If it is loaded from non-writable memory, it will crash and freeze
the machine at the Microsoft logo.

Unsetting this flag will cause GRUB to allocate RWX memory for the
kernel.

Note: This won't fix the stock kernels or the installation image.