pkg/debian: Temporarily add serial suspend
Ubuntu mainline repos are currently stuck at v6.5.7. However, our
patches are for v6.5.8+ and have thus dropped the serial suspend fix,
which was added upstream in v6.5.8. So add it to the Debian/Ubuntu
package specific patches for now so that we can still build
Debian/Ubuntu kernel packages.