pkg/arch: Do not build in serial drivers
Since [1] we should not require the serial drivers to be built in any
more. In essence, the PR causes the SAM driver to return EPROBE_DEFER
when the serial device is not ready.

Specifics on how to best handle this (i.e., which function should
actually return EPROBE_DEFER) are still to be discussed, but for now
this should provide a better workaround that hopefully fixes this
problem as well. So let's test it.

[1]: https://github.com/linux-surface/kernel/pull/152